tts/requirements.txt
chatterbox-tts
fastapi
uvicorn[standard]
torch
torchaudio
python-multipart
